adds concept of SymHash for Types, adds preliminary structs for Modules
Started to implement some very naive "type checking" and realized
that Symbol Ids were only the second half of the story as SymbolIds
would only be generated and there was no way to query for
the SymbolId of a given type.

This came up because of the way generic types are implemented.
Applying TypeArguments to a Type is similar to applying Arugments
to a Function. Every application results in another Type but unlike
primitive and reference values, the equivalence of Types needs to be
evaluated by the compiler. Applying the type of String to Option via
Option<<String>> would return a new Type with the Option's
TypeParameters bound.

Instead of every application of String to Option<<T>> resulting
in a new SymbolId, the compiler needs to be smart enough to understand
that the same TypeArgument is being applied to the same polymorphic
Type which means that the same SymbolId needs to be used.

This change only adds the plumbing to make this possible. The compiler
will currently register two different occurances of Option<<String>>
as two different SymbolIds and crash.