refactor native functions and adds enum intializations
Enums that take contain some type work but enums that have no containing
type break because currently `SomeEnum::SomeVariant` is accesing a shadow
module named `SomeEnum` and getting the NativeFunction `SomeVariant`.

When there is a contained type, this function is actually applied
but when the varient doesn't contain anything, the NativeFunction is
returned without ever applying it.

I'll need to change the Enum struct to have each variant point to
either a NativeFunction OR just an already constructed EnumInstance.
Then when accessing the module `SomeEnum` it will just return an already
constructed EnumInstance rather than a NativeFunction